requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
easyocr
pillow
pandas
openpyxl

